+++ examples/account-selector.js
@@ -1,94 +1,48 @@
 import React from 'react';
-import { render } from 'react-dom';
+import { accounts as accountArray } from './example-data';
 import AccountSelector from '../src/account-selector';
 
-require('./example.less');
+const AccountSelectorExample = () => {
 
-const heading = document.createElement('h1');
-heading.textContent = 'Account selector';
-heading.className = 'ffe-h1';
-document.body.appendChild(heading);
+  const preSelectedAccountNumber = '42010205683';
 
-const accountSelectorDOM = document.createElement('section');
-document.body.appendChild(accountSelectorDOM);
+  function getAccount(accounts, accountNumber) {
+    return accounts.find(account => account.accountNumber === accountNumber);
+  }
 
-const accounts = [
-  {
-    name: 'Skattetrekkskonto',
-    accountNumber: '42010205683',
-    balance: 66546.99,
-    currencyCode: 'USD',
-  },
-  {
-    name: 'Matvererkonto',
-    accountNumber: '42011841044',
-    balance: 24566546.21,
-  },
-  {
-    name: 'Something else konto',
-    accountNumber: '42000186044',
-    balance: 12566546.34,
-    currencyCode: 'EUR',
-  },
-  {
-    name: 'Et virkelig langt kontonavn',
-    accountNumber: '42020382079',
-    balance: 13466546.15,
-  },
-  {
-    name: 'Lønnskonto',
-    accountNumber: '42020382117',
-    balance: 4566546,
-  },
-  {
-    name: 'Martins konto',
-    accountNumber: '42003346573',
-    balance: 0,
-  },
-  {
-    name: 'Terjes konto',
-    accountNumber: '42022319331',
-    balance: 999999.99,
-  },
-  {
-    name: 'Apple Inc',
-    accountNumber: '42000619307',
-    balance: 9999997643.50,
-  },
-];
+  function onChange(accountNumberOrUserInput) {
+    console.log('account changed', accountNumberOrUserInput);
+  }
 
-const value = '42010205683';
+  function onBlur(selectedAccount, inputvalue) {
+    console.log('account selector lost focus, selected account:', selectedAccount);
+  }
 
-function getAccount(accounts, accountNumber) {
-  return accounts.find(account => account.accountNumber === accountNumber);
-}
+  function onFocus() {
+    console.log('account selector gained focus');
+  }
 
-function onChange(accountNumberOrUserInput) {
-  console.log('account changed', accountNumberOrUserInput);
-}
+  function onAccountSelected(value) {
+    console.log('account selected', value);
+  }
 
-function onBlur(value) {
-  console.log(`account selector lost focus, account number or input value: ${value}`);
-}
+  return (
+    <div className="selector">
+      <label htmlFor="custom-id" className="ffe-form-label selector-label">Velg konto</label>
+      <AccountSelector
+        accounts={ accountArray }
+        onAccountSelected={ onAccountSelected }
+        onChange={ onChange }
+        onBlur={ onBlur }
+        onFocus={ onFocus }
+        locale="nb"
+        placeholder="Velg Konto"
+        noMatches="Ingen kontoer"
+        id="custom-id"
+        selectedAccount={ getAccount(accountArray, preSelectedAccountNumber) }
+      />
+    </div>
+  );
+};
 
-function onFocus() {
-  console.log('account selector gained focus');
-}
-
-render(
-  <div>
-    <label htmlFor="custom-id" className="ffe-form-label">Velg konto</label>
-    <AccountSelector
-      accounts={ accounts }
-      onChange={ onChange }
-      onBlur={ onBlur }
-      onFocus={ onFocus }
-      locale="nb"
-      placeholder="Velg Konto"
-      noMatches="Ingen kontoer"
-      id="custom-id"
-      value={  value ? getAccount(accounts, value).name : '' }
-      selectedAccount={ getAccount(accounts, value) }
-    />
-  </div>
-  , accountSelectorDOM);
+export default AccountSelectorExample;