+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 #Changelog
 
+# v3.0.0
+* Major underlying change in structure of code.
+* Add ability to more easily create custom selectors
+* Add a new multi-account selector
+
+# v2.1.5
+* Bugfix: Bundle css from react-scrollbar instead of depending on it. Depending on relative dependencies does not work when using npm@3 og yarn.
+
+# v2.1.4
+* Bugfix: Give input field focus on reset button click
+
 # v2.1.3
 * Bugfix: Fixed various bugs related to race conditions between events
 
@@ -7,7 +18,7 @@
 * Bugfix: make noMatches-field persist when clicked
 * Bugfix: changed noMatches-text to be wrapped in `span` instead of `anchor` to avoid influences of external styles
 
-## v2.1.1 
+## v2.1.1
 * Bugfix: Input field no longer steals focus onBlur
 
 ## v2.1.0